+++ trunk/open-ig/src/hu/openig/scripting/EmptyScripting.java
@@ -351,11 +351,11 @@ public class EmptyScripting implements GameScripting {
 		return true;
 	}
 	@Override
-	public int playerPopulationGrowthOverride(Planet planet, int simulatorValue) {
+	public double playerPopulationGrowthOverride(Planet planet, double simulatorValue) {
 		return simulatorValue;
 	}
 	@Override
-	public int playerTaxIncomeOverride(Planet planet, int simulatorValue) {
+	public double playerTaxIncomeOverride(Planet planet, double simulatorValue) {
 		return simulatorValue;
 	}
 	@Override
